requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
bcrypt==4.0.1
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.7
click==8.3.2
cryptography==46.0.7
ecdsa==0.19.2
fastapi==0.136.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
passlib==1.7.4
pyasn1==0.6.3
pycparser==3.0
pydantic==2.13.3
pydantic_core==2.46.3
pylast==7.0.2
python-dotenv==1.2.2
python-jose==3.5.0
redis==7.4.0
requests==2.33.1
rsa==4.9.1
six==1.17.0
spotipy==2.26.0
SQLAlchemy==2.0.49
starlette==1.0.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uvicorn==0.44.0
